Addressing the comments. Doing fixes and reverts.
