--- Core/Goals/TargetFinder/ITargetFinder.cs
@@ -1,13 +1,0 @@
-﻿using SharedLib.NpcFinder;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Core.Goals
-{
-    public interface ITargetFinder
-    {
-        Task<bool> Search(string source, CancellationToken cancellationToken);
-    }
-}
